MFAH/requirements.txt
asgiref==3.5.2
Django==4.0.4
django-colorfield==0.7.0
django-jazzmin==2.5.0
Pillow==9.1.1
sqlparse==0.4.2
tzdata==2022.1
